Create Static Components
Create a JavaScript file and a CSS file for each of six components in the Jammming app.
The components are: App, SearchBar, SearchResults, Playlist, TrackList, Track.
Inside Track component add renderAcion() method  that displays
a - anchor tag if the isRemoval property is true,
and a + anchor tag if the isRemoval property is false.